# src/test/java/com/poovarasan/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "see https://github.com/StefanSchubert/sabi",
        "version": "1.0",
        "title": "REST API documentation",
        "contact": {
            "name": "Poovarasan Vasudevan",
            "url": "http://shpt.in",
            "email": "poosan9@gmail.com"
        },
        "license": {
            "name": "MIT Licence (MIT)",
            "url": "https://github.com/StefanSchubert/sabi/blob/master/LICENSE"
        }
    },
    "host": "localhost:8097",
    "basePath": "/",
    "tags": [
        {
            "name": "rest-controller",
            "description": "Rest Controller"
        }
    ],
    "paths": {
        "/api/me": {
            "get": {
                "tags": [
                    "rest-controller"
                ],
                "summary": "me",
                "operationId": "meUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/role/{userid}": {
            "get": {
                "tags": [
                    "rest-controller"
                ],
                "summary": "userRole",
                "operationId": "userRoleUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "userid",
                        "in": "path",
                        "description": "userid",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Role"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/socket/ping": {
            "get": {
                "tags": [
                    "rest-controller"
                ],
                "summary": "pingSocket",
                "operationId": "pingSocketUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "rest-controller"
                ],
                "summary": "getAllUser",
                "operationId": "getAllUserUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "definitions": {
        "Role": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "User": {
            "type": "object",
            "properties": {
                "createdDate": {
                    "$ref": "#/definitions/Timestamp"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "updatedDate": {
                    "$ref": "#/definitions/Timestamp"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "Timestamp": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "integer",
                    "format": "int32"
                },
                "day": {
                    "type": "integer",
                    "format": "int32"
                },
                "hours": {
                    "type": "integer",
                    "format": "int32"
                },
                "minutes": {
                    "type": "integer",
                    "format": "int32"
                },
                "month": {
                    "type": "integer",
                    "format": "int32"
                },
                "nanos": {
                    "type": "integer",
                    "format": "int32"
                },
                "seconds": {
                    "type": "integer",
                    "format": "int32"
                },
                "time": {
                    "type": "integer",
                    "format": "int64"
                },
                "timezoneOffset": {
                    "type": "integer",
                    "format": "int32"
                },
                "year": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        }
    }

}